Marisol - Created RegisterUser Page
Hats off to Marisol for the creation of this RegisterUser page. Still needs tweaking regarding the placement of some of the stack panels and docks. However, as a temporary fix, I was able to tweak some of the opacity values for the values so that the input boxes would be more visible to the user. Foundation has been set, just needs to be perfected.