Completed Startup Navigation
Completed QRbert Startup Navigation
I set the startup URI in the app.xaml file to be WelcomeScreen, however, for some reason, it doesn't load, it goes straight to the GetStarted window. For the time being this is fine. The button Get Started on that window takes you to the MainWindow. Also updated namespaces and added Scroll Bars for some windows.