Started Registration Navigation
Started Registration Navigation; updated App file to force StartupUri to be the GetStarted Window. This will have to be updated eventually in a main pull request in the future. Used the NavigationService, which is a Windows class to support Navigation between Windows and Pages. So far, this has not worked. Lastly, updated Register.xaml to call the RegisterUser function that Denise created, rather than Jonathan's temporary function to view the password to ensure the correct functionality of all textboxes and functions. This is also something that will have to updated to stay in a future pull request for main. Committing and pushing this code so I can continue to work on this when I have time at work.