Switcher class for staff/volunteer portal navigation
Quickly implemented a switcher class I found online that allows us to switch and navigate rather than change the content of the window/page.
Changed several files to add the button clicks, methods, and events for each of the pages
Changed some files to be windows instead of pages so that the switcher class could do its job easier
Most of my functions have comments

I ask that everyone take a look at these file changes and implement what you need to, e.g. Denise: database queries to log in a user
David: given a query, compare the inputted password as a hash with the one that you generate and save in the database
Marisol: menu item replacement, see comments in the code for staff and volunteer portal
Jonathan: your own pages that I already assigned to you