Camera classes
I added the necessary classes to read an image from a file.
For now, I have the access to the camera from the Forgot your password button in the LogIn_Register.xaml file.
Essentially, I create an object within this button handler and open the camera. Currently, I not getting errors but it does crash when trying to open the camera. Have to figure out why.