Adds functionality for the camera
The camera is now working. I made a QRCodeScanner class that grabs an image from the webcam and displays the next continuously. This class also takes care of decoding.
I added the correct navigation of buttons in the respective files such as GetStarted and LogIn to make sure the navigation is correct. I added code to the Register c# file to allow the user to take a picture of their QR code which they will need to Log In after being redirected. I need SQL queries to verify the user. For now, I have made the QR code not a random sequence of characters, but this can be changed. Lastly, redirect them to the Staff Portal.