Created pet report page
The staff will be navigated to this page once the staff user has entered the pet id in the view pet report page that they wish to view. The page is missing a stack panel that will make a pdf visible as well as a button that allows the user to download it.