Update button clicks for viewing report
The user can now click the Scan and View button and will be prompted to scan the Pet's QR code in a new window
I need a query to authenticate the QR code. After this, a new window to view the PDF is opened.
Things to work now: if statement that verifies the QR code decoded string result...