Updated 98% of current windows for navigation
Added functionality for almost all windows, current code is for windows that are assignable given information or have been implemented in other windows. For now, the Staff Portal is functionable between all its StaffX windows as well as the Volunteer Windows. I created, edited, and modified some windows in order to keep on par. Sorry Marisol, I had to. I also deleted some files that we no longer need. I have edited David's QR Code Generator code and moved it to the Scanner File, makes more sense as well as edited the functionality in the Register Window to use his code.   I renamed a couple of files as well to make sure they are consistent with our naming convention.