Pet authentication and cleanup
Added code to authenticate a Pet's QR Code and cleanup some files. Also removed the entire ZBar libraries, dependencies... the entire codebase. I tried to organize the files into different folders, however, I was getting namespace conflicts and it would have been too much to reorganize and format the code. Will have to leave that way. Camera is working for me. I added a public string for the DB connection in the Switcher class. To use this variable, just call Switcher.connectionString
I edited all the dependent window classes for the connection string. Going forward, just do what I said above. As mentioned, scanning the Pet's QR code seems to work, but Denise brought to light a problem with the camera on her branch when registering. Will have to pull the next changes because it doesn't seem to be a problem for me.