More navigation, including MyPet support
Added more navigation for new buttons and windows created. Untested most of this code as I have not created Pets, however, queries were made to get relevant Pet and save the current user and Pet scanned into the Switcher class. You can access the connection string from anywhere in this class, as well as the current user and petID scanned. Worked on the TimeSheet windows, still waiting on Marisol to make them, but they should be simple enough to add the Click methods to them.