Change email feature
Staff can now change their email. Added support for user to input their email. Since email is saved, the code queries the relevant info to update the email in the Registration table. Don't know where else it has to be updated, but the query can be copied if need be for the other tables. It then takes the user back to the MyAccount page.