Added support for bell icon
Switcher class has boolean to save state of whether there is a Pet that has been neglected and also saves the query. The Bell Icon is invisible unless the boolean is true. This can accessed globally. If the boolean is false, the method does nothing. Otherwise, it redirects them to the Neglected Pets Window